requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

nemo_toolkit[asr]==2.4.0

soundfile==0.13.1
numpy==1.26.4
